+++ xrspatial/multispectral.py
@@ -11,6 +11,7 @@ from xarray import DataArray
 
 from xrspatial.utils import (ArrayTypeFunctionMapping, cuda_args, ngjit, not_implemented_func,
                              validate_arrays)
+from xrspatial.dataset_support import supports_dataset_bands
 
 # 3rd-party
 try:
@@ -75,6 +76,7 @@ def _arvi_dask_cupy(nir_data, red_data, blue_data):
     return out
 
 
+@supports_dataset_bands(nir='nir_agg', red='red_agg', blue='blue_agg')
 def arvi(nir_agg: xr.DataArray,
          red_agg: xr.DataArray,
          blue_agg: xr.DataArray,
@@ -94,6 +96,12 @@ def arvi(nir_agg: xr.DataArray,
         2D array of blue band data.
     name : str, default='arvi'
         Name of output DataArray.
+
+    Alternatively, a single ``xr.Dataset`` may be passed as the first
+    argument with keyword arguments mapping band names to Dataset
+    variables. For example::
+
+        arvi(ds, nir='B8', red='B4', blue='B2')
 
     Returns
     -------
@@ -215,6 +223,7 @@ def _evi_dask_cupy(nir_data, red_data, blue_data, c1, c2, soil_factor, gain):
     return out
 
 
+@supports_dataset_bands(nir='nir_agg', red='red_agg', blue='blue_agg')
 def evi(nir_agg: xr.DataArray,
         red_agg: xr.DataArray,
         blue_agg: xr.DataArray,
@@ -246,6 +255,12 @@ def evi(nir_agg: xr.DataArray,
         Amplitude adjustment factor.
     name : str, default='evi'
         Name of output DataArray.
+
+    Alternatively, a single ``xr.Dataset`` may be passed as the first
+    argument with keyword arguments mapping band names to Dataset
+    variables. For example::
+
+        evi(ds, nir='B8', red='B4', blue='B2')
 
     Returns
     -------
@@ -374,6 +389,7 @@ def _gci_dask_cupy(nir_data, green_data):
     return out
 
 
+@supports_dataset_bands(nir='nir_agg', green='green_agg')
 def gci(nir_agg: xr.DataArray,
         green_agg: xr.DataArray,
         name='gci'):
@@ -390,6 +406,12 @@ def gci(nir_agg: xr.DataArray,
         2D array of green band data.
     name : str, default='gci'
         Name of output DataArray.
+
+    Alternatively, a single ``xr.Dataset`` may be passed as the first
+    argument with keyword arguments mapping band names to Dataset
+    variables. For example::
+
+        gci(ds, nir='B8', green='B3')
 
     Returns
     -------
@@ -451,6 +473,7 @@ def gci(nir_agg: xr.DataArray,
 
 
 # NBR ----------
+@supports_dataset_bands(nir='nir_agg', swir2='swir2_agg')
 def nbr(nir_agg: xr.DataArray,
         swir2_agg: xr.DataArray,
         name='nbr'):
@@ -468,6 +491,12 @@ def nbr(nir_agg: xr.DataArray,
         (Landsat 8: Band 7)
     name : str, default='nbr'
         Name of output DataArray.
+
+    Alternatively, a single ``xr.Dataset`` may be passed as the first
+    argument with keyword arguments mapping band names to Dataset
+    variables. For example::
+
+        nbr(ds, nir='B8', swir2='B12')
 
     Returns
     -------
@@ -529,6 +558,7 @@ def nbr(nir_agg: xr.DataArray,
                      attrs=nir_agg.attrs)
 
 
+@supports_dataset_bands(swir1='swir1_agg', swir2='swir2_agg')
 def nbr2(swir1_agg: xr.DataArray,
          swir2_agg: xr.DataArray,
          name='nbr2'):
@@ -551,6 +581,12 @@ def nbr2(swir1_agg: xr.DataArray,
         (Landsat 8: Band 7)
     name : str default='nbr2'
         Name of output DataArray.
+
+    Alternatively, a single ``xr.Dataset`` may be passed as the first
+    argument with keyword arguments mapping band names to Dataset
+    variables. For example::
+
+        nbr2(ds, swir1='B11', swir2='B12')
 
     Returns
     -------
@@ -614,6 +650,7 @@ def nbr2(swir1_agg: xr.DataArray,
 
 
 # NDVI ----------
+@supports_dataset_bands(nir='nir_agg', red='red_agg')
 def ndvi(nir_agg: xr.DataArray,
          red_agg: xr.DataArray,
          name='ndvi'):
@@ -629,6 +666,12 @@ def ndvi(nir_agg: xr.DataArray,
         2D array red band data.
     name : str default='ndvi'
         Name of output DataArray.
+
+    Alternatively, a single ``xr.Dataset`` may be passed as the first
+    argument with keyword arguments mapping band names to Dataset
+    variables. For example::
+
+        ndvi(ds, nir='B8', red='B4')
 
     Returns
     -------
@@ -691,6 +734,7 @@ def ndvi(nir_agg: xr.DataArray,
 
 
 # NDMI ----------
+@supports_dataset_bands(nir='nir_agg', swir1='swir1_agg')
 def ndmi(nir_agg: xr.DataArray,
          swir1_agg: xr.DataArray,
          name='ndmi'):
@@ -710,6 +754,12 @@ def ndmi(nir_agg: xr.DataArray,
         (Landsat 8: Band 6)
     name: str, default='ndmi'
         Name of output DataArray.
+
+    Alternatively, a single ``xr.Dataset`` may be passed as the first
+    argument with keyword arguments mapping band names to Dataset
+    variables. For example::
+
+        ndmi(ds, nir='B8', swir1='B11')
 
     Returns
     -------
@@ -874,6 +924,7 @@ def _savi_dask_cupy(nir_data, red_data, soil_factor):
 
 
 # SAVI ----------
+@supports_dataset_bands(nir='nir_agg', red='red_agg')
 def savi(nir_agg: xr.DataArray,
          red_agg: xr.DataArray,
          soil_factor: float = 1.0,
@@ -894,6 +945,12 @@ def savi(nir_agg: xr.DataArray,
         When set to zero, savi will return the same as ndvi.
     name : str, default='savi'
         Name of output DataArray.
+
+    Alternatively, a single ``xr.Dataset`` may be passed as the first
+    argument with keyword arguments mapping band names to Dataset
+    variables. For example::
+
+        savi(ds, nir='B8', red='B4')
 
     Returns
     -------
@@ -1006,6 +1063,7 @@ def _sipi_dask_cupy(nir_data, red_data, blue_data):
     return out
 
 
+@supports_dataset_bands(nir='nir_agg', red='red_agg', blue='blue_agg')
 def sipi(nir_agg: xr.DataArray,
          red_agg: xr.DataArray,
          blue_agg: xr.DataArray,
@@ -1024,6 +1082,12 @@ def sipi(nir_agg: xr.DataArray,
         2D array of blue band data.
     name: str, default='sipi'
         Name of output DataArray.
+
+    Alternatively, a single ``xr.Dataset`` may be passed as the first
+    argument with keyword arguments mapping band names to Dataset
+    variables. For example::
+
+        sipi(ds, nir='B8', red='B4', blue='B2')
 
     Returns
     -------
@@ -1142,6 +1206,7 @@ def _ebbi_dask_cupy(red_data, swir_data, tir_data):
     return out
 
 
+@supports_dataset_bands(red='red_agg', swir='swir_agg', tir='tir_agg')
 def ebbi(red_agg: xr.DataArray,
          swir_agg: xr.DataArray,
          tir_agg: xr.DataArray,
@@ -1160,6 +1225,12 @@ def ebbi(red_agg: xr.DataArray,
         2D array of thermal infrared band data.
     name: str, default='ebbi'
         Name of output DataArray.
+
+    Alternatively, a single ``xr.Dataset`` may be passed as the first
+    argument with keyword arguments mapping band names to Dataset
+    variables. For example::
+
+        ebbi(ds, red='B4', swir='B11', tir='B10')
 
     Returns
     -------